apps/api/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async Support
httpx>=0.26.0
aiohttp>=3.9.0

# Pydantic for Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# Supabase
supabase>=2.3.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# AI/ML
google-generativeai>=0.3.2
openai>=1.10.0

# Image Processing
pillow>=10.2.0
python-pptx>=0.6.23

# Web Scraping (for Brand DNA)
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# External APIs
requests>=2.31.0
